rest_framework_cassandra_engine/tests/requirements.txt
argparse==1.2.1
djangorestframework==3.0.1
nose==1.3.3
Django==1.7.1
blist==1.3.6
cassandra-driver==2.1.2
python-dateutil==2.3
python-mimeparse==0.1.4
cqlengine==0.20.0
-e git+https://github.com/talus-marks/djangotoolbox.git#egg=djangotoolbox
-e git+https://github.com/talus-marks/django-cassandra-engine.git#egg=django-cassandra-engine
futures==2.2.0
six==1.8.0
